feat(test): Add var to run tests on real flintlock (#44)
FFS I managed to put quite a lot in here which I forgot about and now
cannot find the seams to split up.

- As part of this I have heavily refactored the integration tests to make
  them only test core features. As we refactor, we will add granular unit
  tests which will check against all the various flags and config options.

- I fixed the fake flintlock server as it actually did more that
  flintlock.

- I have also fixed a bug in which the command hangs if the server cannot
  be connected to (removed `grpc.WithBlock()`).

- The `grpc-port` flag is gone, now it is at one with the
  `grpc-address`. This mirrors what flintlock does.

- I added a `name` flag to `list` so now you can list by name/ns
  groupings.

- I updated example.json to be in line with the flintlock api.

+++ test/integration/integration_suite_test.go
@@ -1,6 +1,8 @@
 package integration_test
 
 import (
+	"fmt"
+	"os"
 	"os/exec"
 	"testing"
 
@@ -12,7 +14,7 @@ import (
 
 var (
 	cliBin  string
-	address = "127.0.0.1"
+	address string
 )
 
 func TestIntegration(t *testing.T) {
@@ -30,20 +32,30 @@ func TestIntegration(t *testing.T) {
 		serverBin, err = gexec.Build("github.com/warehouse-13/hammertime/test/fakeserver")
 		Expect(err).NotTo(HaveOccurred())
 
-		serverCmd := exec.Command(serverBin)
-		serverSession, err = gexec.Start(serverCmd, GinkgoWriter, GinkgoWriter)
-		Expect(err).NotTo(HaveOccurred())
+		if remote_test_server := os.Getenv("TEST_SERVER"); remote_test_server != "" {
+			address = remote_test_server
+			fmt.Fprintf(GinkgoWriter, "Using real Flintlock server at %s: tests may take a little longer", address)
+		} else {
+			serverCmd := exec.Command(serverBin)
+			serverSession, err = gexec.Start(serverCmd, GinkgoWriter, GinkgoWriter)
+			Expect(err).NotTo(HaveOccurred())
+
+			address = "127.0.0.1:9090"
+		}
 
 		// Sometimes the server doesn't start immediately, so we check that an
 		// endpoint is reachable before we carry on with the test.
 		getCmd := command{action: "get", args: []string{"--id", "foobar"}}
 		getSession := executeCommand(getCmd)
 		Eventually(getSession, "10s").Should(gexec.Exit(1))
-		Eventually(getSession.Err).Should(gbytes.Say("OHH WHAT A DISASTER"))
+		Eventually(getSession.Err).Should(gbytes.Say("rpc error"))
 	})
 
 	AfterSuite(func() {
-		serverSession.Terminate().Wait()
+		if serverSession != nil {
+			serverSession.Terminate().Wait()
+		}
+
 		gexec.CleanupBuildArtifacts()
 	})
